rename string folder as text
