+++ src/text/index.ts
