+++ src/text/ltrim.ts
