+++ src/text/rtrim.ts
