+++ src/text/test/ltrim-cases.ts
