+++ src/text/test/ltrim.spec.ts
