+++ src/text/test/rtrim-cases.ts
