+++ src/text/test/rtrim.spec.ts
