+++ src/text/test/trim-cases.ts
