+++ src/text/test/trim.spec.ts
