+++ src/text/trim.ts
